+++ NEWS.md
@@ -3,6 +3,7 @@ Unreleased
 * Remove `suspenders` system executable
 * Introduce `suspenders:accessibility` generator
 * Introduce `Suspenders::Generators::APIAppUnsupported` module and concern
+* Introduce `suspenders:lint` generator
 
 20230113.0 (January, 13, 2023)